add tests for unique mark

+++ src/unique-mark.ts
@@ -1,14 +1,15 @@
 #!/usr/bin/env node
+/* eslint-disable import/exports-last */
 import { execSync } from 'child_process'
 import { existsSync, readFileSync, writeFileSync } from 'fs'
 import path from 'path'
 import glob from 'tiny-glob'
-import { blue } from './colors'
-import { nbThird } from './constants'
-import { formatDate } from './dates'
-import { injectMark, parseJson } from './strings'
+import { blue } from './colors.js'
+import { formatDate } from './dates.js'
+import { injectMark, parseJson } from './strings.js'
 import type { PackageJson } from './types'
 
+/* c8 ignore next 9 */
 /**
  * Log a message to the console
  * @param message the message to log
@@ -21,26 +22,23 @@ function log (message: string, value = '') {
 
 /**
  * Load the package.json file data
+ * @param location the location of the package.json file
  * @returns {string} the package.json version
  */
-function getPackageJsonVersion () {
-  const pkgLocation = path.join(process.cwd(), 'package.json')
-  if (!existsSync(pkgLocation)) throw new Error(`package.json was not found in ${pkgLocation}, aborting.`)
-  const content = readFileSync(pkgLocation, 'utf8')
+export function getPackageJsonVersion (location = path.join(process.cwd(), 'package.json')) {
+  if (!existsSync(location)) throw new Error(`package.json was not found in ${location}, aborting.`)
+  const content = readFileSync(location, 'utf8')
   const { error, value } = parseJson<PackageJson>(content)
-  if (error) throw new Error(`package.json in ${pkgLocation} is not a valid JSON, aborting.`)
+  if (error) throw new Error(`package.json in ${location} is not a valid JSON, aborting.`)
   return value.version
 }
 
 /**
  * Get the files to inject the mark in
+ * @param target the glob to get the files from, like "public/index.html" or "public/*.js"
  * @returns {Promise<string[]>} the files to inject the mark in
  */
-async function getTargetFiles () {
-  const expectedNbArguments = 3
-  const hasTargetSpecified = process.argv.length === expectedNbArguments
-  const defaultTarget = 'public/index.html'
-  const target = (hasTargetSpecified && typeof process.argv[nbThird] === 'string') ? (process.argv[nbThird] ?? defaultTarget) : defaultTarget
+export async function getTargetFiles (target = process.argv[2] ?? 'public/index.html') { // eslint-disable-line @typescript-eslint/no-magic-numbers
   const files = await glob(target)
   if (files.length === 0) throw new Error(`no file found for target "${target}", aborting.`)
   return files
@@ -54,8 +52,9 @@ async function getTargetFiles () {
  * @param root0.version the version to use, if empty, will use the version from package.json
  * @returns {string} the mark to inject, like "4.2.0 - 123abc45 - 01/01/2021 12:00:00"
  */
-function generateMark ({ commit = '', date = formatDate(new Date(), 'dd/MM/yyyy HH:mm:ss'), version = '' }) {
+export function generateMark ({ commit = '', date = formatDate(new Date(), 'dd/MM/yyyy HH:mm:ss'), version = '' }) {
   let finalCommit = commit
+  /* c8 ignore next */
   if (commit === '') finalCommit = execSync('git rev-parse --short HEAD', { cwd: process.cwd() }).toString().trim()
   return `${version} - ${finalCommit} - ${date}`
 }
@@ -67,23 +66,27 @@ function generateMark ({ commit = '', date = formatDate(new Date(), 'dd/MM/yyyy 
  * @param root0.placeholder the placeholder to use to inject the mark
  * @param root0.mark the mark to inject
  * @param root0.files the files to inject the mark in
+ * @param root0.isReadOnly if true, will not write the files to disk (useful for unit tests)
  * @returns {number} the total amount of mark injection in the targeted files
  */
-function injectMarkInFiles ({ placeholder = 'unique-mark', mark = 'no-mark', files = [] }: { placeholder?: string; mark?: string; files?: string[] }) {
+export function injectMarkInFiles ({ placeholder = 'unique-mark', mark = 'no-mark', files = [], isReadOnly = false }: { placeholder?: string; mark?: string; files?: string[]; isReadOnly?: boolean }) {
   let totalInjections = 0
+  const logs: string[] = []
+  const markRegex = new RegExp(mark, 'gu') // eslint-disable-line security/detect-non-literal-regexp
   files.forEach(file => {
     const content = readFileSync(file, 'utf8')
     if (!content.includes(placeholder)) throw new Error(`could not find a place to inject in ${file}, aborting.\n\nPlease use one or more of these placeholders :  <span id="${placeholder}"></span>  <meta name="${placeholder}" content="">  __${placeholder}__`)
     const updatedContent = injectMark(content, placeholder, mark)
-    // eslint-disable-next-line security/detect-non-literal-regexp, regexp/prefer-regexp-exec
-    const times = (updatedContent.match(new RegExp(mark, 'gu')) ?? []).length
-    writeFileSync(file, updatedContent)
-    log(`injected in ${file}`, `${times} time${times > 1 ? 's' : ''}\n`)
+    const times = updatedContent.match(markRegex)?.length /* c8 ignore next */ ?? 0 // eslint-disable-line regexp/prefer-regexp-exec
+    /* c8 ignore next 2 */
+    if (!isReadOnly) writeFileSync(file, updatedContent)
+    logs.push(`injected in ${file} : ${times} time${times > 1 ? 's' : ''}`)
     totalInjections += times
   })
-  return totalInjections
+  return { totalInjections, logs }
 }
 
+/* c8 ignore next 19 */
 /**
  * Main function
  */
@@ -95,8 +98,12 @@ async function init () {
   if (isVerbose) log(`found ${files.length} file${files.length > 1 ? 's' : ''} to inject mark`, files.join(', '))
   const mark = generateMark({ version })
   if (isVerbose) log('generated mark', mark)
-  injectMarkInFiles({ mark, files })
+  const { logs, totalInjections } = injectMarkInFiles({ mark, files })
+  // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+  logs.forEach(line => { log(...(line.split(':') as [string, string])) })
+  log('Total injections', String(totalInjections))
 }
 
 // eslint-disable-next-line unicorn/prefer-top-level-await
-void init()
+if (process.env.NODE_ENV !== 'test') void init()
+
